refactor: move Table to a separate component
